// src/types/form.tsx
export interface FormData {
    name: string;
    nationalId: string;
    dob: string;
    gender: string;
    address: string;
    city: string;
    state: string;
    country: string;
    phone: string;
    email: string;
    maritalStatus: string;
    dependents: string;
    employmentStatus: string;
    monthlyIncome: string;
    housingStatus: string;
    financialSituation: string;
    employmentCircumstances: string;
    reasonForApplying: string;
}